#2 The number of frames that should be generated is now passed to the backend.  close #2.

+++ PathfinderPrototype/Assets/Editor/AnimationGeneratorEditor.cs
@@ -144,7 +144,7 @@ public class AnimationGeneratorEditor : Editor
 
         GUILayout.BeginHorizontal();
         GUILayout.Label("Frames of Animation");
-        generator.framesOfAnimation = CustomGUILayout.FloatField(generator.framesOfAnimation);
+        generator.framesOfAnimation = (int)CustomGUILayout.FloatField((float)generator.framesOfAnimation);
         GUILayout.EndHorizontal();
         GUILayout.BeginHorizontal();
